xml加载work可输出加载记录
