+++ src/test/resources/application.properties
@@ -1,6 +1,12 @@
+#当指定的work、taskbean、task不存在时忽略，程序继续执行
 taskflow.ignoreNoExists=true
+#开启运行时动态重载work
 taskflow.reload.enable=true
+#开启task执行记录；为true时每个task执行会记录下workcontext的快照
 taskflow.work.traceable=false
+#同步task等待异步task传参时的等待时间；单位：毫秒
 taskflow.task.asyncTimeOut=5000
+#开启打印work注册记录
 taskflow.log.printable=true
+#打印work注册记录时是否输出详情
 taskflow.log.printDetail=false
